# src/main/java/ua/com/foxminded/university/controller/web/thymeleaf/dto/output/StudentDtoOut.java
package ua.com.foxminded.university.controller.web.thymeleaf.dto.output;

import lombok.AllArgsConstructor;
import lombok.EqualsAndHashCode;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;
import lombok.ToString;
import lombok.experimental.SuperBuilder;

@SuperBuilder
@Setter
@Getter
@AllArgsConstructor
@NoArgsConstructor
@ToString(callSuper = true)
@EqualsAndHashCode(callSuper = true)
public class StudentDtoOut extends PersonDtoOut {
    private GroupDtoOut group;
}