# src/main/resources/swagger.yaml
swagger: '2.0'
info:
  description: Sample API for University tutorial
  version: '1.0'
  title: University API
  termsOfService: Free to use
  contact:
    name: Yaroslav Voronovskyi
    url: https://www.linkedin.com/in/yaroslav-voronovskyi-b8a5a3191/
    email: yaroslav.voronovskyi@gmail.com
  license:
    name: API License
    url: http://localhost:8081/
host: localhost:8081
basePath: /
...
paths:
...
  /rest/timetable:
    get:
      tags:
        - time-table-rest-controller
      summary: getAllTimeTables
      operationId: getAllTimeTablesUsingGET
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/TimeTableDtoOut'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    post:
      tags:
        - time-table-rest-controller
      summary: createTimeTable
      operationId: createTimeTableUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: timeTableDtoIn
          description: timeTableDtoIn
          required: true
          schema:
            $ref: '#/definitions/TimeTableDtoIn'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/TimeTableDtoOut'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
definitions:
...
  AuditoryDtoIn:
    type: object
    properties:
      capacity:
        type: integer
        format: int32
        minimum: 20
        maximum: 50
        exclusiveMinimum: false
        exclusiveMaximum: false
      id:
        type: integer
        format: int32
    title: AuditoryDtoIn
  AuditoryDtoOut:
    type: object
    properties:
      capacity:
        type: integer
        format: int32
      id:
        type: integer
        format: int32
    title: AuditoryDtoOut
  CourseDtoIn:
    type: object
    required:
      - name
    properties:
      groupList:
        type: array
        items:
          $ref: '#/definitions/GroupDtoIn'
      id:
        type: integer
        format: int32
      name:
        type: string
        minLength: 3
        maxLength: 10
      subject:
        $ref: '#/definitions/SubjectDtoIn'
      timeTableList:
        type: array
        items:
          $ref: '#/definitions/TimeTableDtoIn'
    title: CourseDtoIn
  CourseDtoOut:
    type: object
    properties:
      groupList:
        type: array
        items:
          $ref: '#/definitions/GroupDtoOut'
      id:
        type: integer
        format: int32
      name:
        type: string
      subject:
        $ref: '#/definitions/SubjectDtoOut'
      timeTableList:
        type: array
        items:
          $ref: '#/definitions/TimeTableDtoOut'
    title: CourseDtoOut
  GroupDtoIn:
    type: object
    required:
      - name
    properties:
      courseList:
        type: array
        items:
          $ref: '#/definitions/CourseDtoIn'
      id:
        type: integer
        format: int32
      name:
        type: string
        minLength: 5
        maxLength: 5
      studentList:
        type: array
        items:
          $ref: '#/definitions/StudentDtoIn'
    title: GroupDtoIn
  GroupDtoOut:
    type: object
    properties:
      courseList:
        type: array
        items:
          $ref: '#/definitions/CourseDtoOut'
      id:
        type: integer
        format: int32
      name:
        type: string
      studentList:
        type: array
        items:
          $ref: '#/definitions/StudentDtoOut'
    title: GroupDtoOut
  ProfessorDtoIn:
    type: object
    required:
      - firstName
      - lastName
    properties:
      firstName:
        type: string
        minLength: 3
        maxLength: 15
      id:
        type: integer
        format: int32
      lastName:
        type: string
        minLength: 3
        maxLength: 15
      subjectList:
        type: array
        items:
          $ref: '#/definitions/SubjectDtoIn'
    title: ProfessorDtoIn
  ProfessorDtoOut:
    type: object
    properties:
      firstName:
        type: string
      id:
        type: integer
        format: int32
      lastName:
        type: string
      subjectList:
        type: array
        items:
          $ref: '#/definitions/SubjectDtoOut'
    title: ProfessorDtoOut
  StudentDtoIn:
    type: object
    required:
      - firstName
      - lastName
    properties:
      firstName:
        type: string
        minLength: 3
        maxLength: 15
      group:
        $ref: '#/definitions/GroupDtoIn'
      id:
        type: integer
        format: int32
      lastName:
        type: string
        minLength: 3
        maxLength: 15
    title: StudentDtoIn
  StudentDtoOut:
    type: object
    properties:
      firstName:
        type: string
      group:
        $ref: '#/definitions/GroupDtoOut'
      id:
        type: integer
        format: int32
      lastName:
        type: string
    title: StudentDtoOut
  SubjectDtoIn:
    type: object
    required:
      - description
      - name
    properties:
      description:
        type: string
        minLength: 15
        maxLength: 350
      id:
        type: integer
        format: int32
      name:
        type: string
        minLength: 4
        maxLength: 15
      professor:
        $ref: '#/definitions/ProfessorDtoIn'
    title: SubjectDtoIn
  SubjectDtoOut:
    type: object
    properties:
      description:
        type: string
      id:
        type: integer
        format: int32
      name:
        type: string
      professor:
        $ref: '#/definitions/ProfessorDtoOut'
    title: SubjectDtoOut
  TimeTableDtoIn:
    type: object
    properties:
      auditory:
        $ref: '#/definitions/AuditoryDtoIn'
      course:
        $ref: '#/definitions/CourseDtoIn'
      date:
        type: string
        example: yyyy-MM-dd
      endTime:
        type: string
        example: HH:mm:ss
      id:
        type: integer
        format: int32
      startTime:
        type: string
        example: HH:mm:ss
    title: TimeTableDtoIn
  TimeTableDtoOut:
    type: object
    properties:
      auditory:
        $ref: '#/definitions/AuditoryDtoOut'
      course:
        $ref: '#/definitions/CourseDtoOut'
      date:
        type: string
        example: yyyy-MM-dd
      endTime:
        type: string
        example: HH:mm:ss
      id:
        type: integer
        format: int32
      startTime:
        type: string
        example: HH:mm:ss
    title: TimeTableDtoOut